# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BALE BELEQ NEWS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SetupStatus {
  setupRequired: boolean;
}

export interface SetupInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
  rememberMe?: boolean;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  EDITOR: 'EDITOR',
  WARTAWAN: 'WARTAWAN',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  EDITOR: 'EDITOR',
  WARTAWAN: 'WARTAWAN',
} as const;

export interface UserInput {
  name: string;
  email: string;
  password: string;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  EDITOR: 'EDITOR',
  WARTAWAN: 'WARTAWAN',
} as const;

export interface UserUpdate {
  name?: string;
  email?: string;
  /** @nullable */
  password?: string | null;
  role?: UserUpdateRole;
}

export type ArticleStatus = typeof ArticleStatus[keyof typeof ArticleStatus];


export const ArticleStatus = {
  DRAFT: 'DRAFT',
  PENDING_REVIEW: 'PENDING_REVIEW',
  PUBLISHED: 'PUBLISHED',
} as const;

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  articleCount?: number;
}

export interface Article {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  coverImage?: string | null;
  status: ArticleStatus;
  isBreaking?: boolean;
  isFeatured?: boolean;
  authorId: number;
  /** @nullable */
  categoryId?: number | null;
  category?: Category;
  author?: User;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
  viewCount?: number;
}

export interface ArticleList {
  articles: Article[];
  total: number;
  page: number;
  limit: number;
}

export type ArticleInputStatus = typeof ArticleInputStatus[keyof typeof ArticleInputStatus];


export const ArticleInputStatus = {
  DRAFT: 'DRAFT',
  PENDING_REVIEW: 'PENDING_REVIEW',
  PUBLISHED: 'PUBLISHED',
} as const;

export interface ArticleInput {
  title: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  categoryId?: number | null;
  isBreaking?: boolean;
  isFeatured?: boolean;
  status?: ArticleInputStatus;
}

export type ArticleUpdateStatus = typeof ArticleUpdateStatus[keyof typeof ArticleUpdateStatus];


export const ArticleUpdateStatus = {
  DRAFT: 'DRAFT',
  PENDING_REVIEW: 'PENDING_REVIEW',
  PUBLISHED: 'PUBLISHED',
} as const;

export interface ArticleUpdate {
  title?: string;
  /** @nullable */
  excerpt?: string | null;
  content?: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  categoryId?: number | null;
  isBreaking?: boolean;
  isFeatured?: boolean;
  status?: ArticleUpdateStatus;
}

export interface CategoryInput {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface Media {
  id: number;
  filename: string;
  url: string;
  type: string;
  size?: number;
  createdAt: string;
}

export interface MediaInput {
  filename: string;
  url: string;
  type: string;
  size?: number;
}

export interface Settings {
  id: number;
  siteName: string;
  /** @nullable */
  siteSlogan: string | null;
  /** @nullable */
  siteDescription?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  footerText?: string | null;
  /** @nullable */
  socialFacebook?: string | null;
  /** @nullable */
  socialInstagram?: string | null;
  /** @nullable */
  socialTwitter?: string | null;
  /** @nullable */
  socialYoutube?: string | null;
}

export interface SettingsUpdate {
  siteName?: string;
  /** @nullable */
  siteSlogan?: string | null;
  /** @nullable */
  siteDescription?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  footerText?: string | null;
  /** @nullable */
  socialFacebook?: string | null;
  /** @nullable */
  socialInstagram?: string | null;
  /** @nullable */
  socialTwitter?: string | null;
  /** @nullable */
  socialYoutube?: string | null;
}

export interface DashboardStats {
  totalArticles: number;
  publishedArticles: number;
  draftArticles: number;
  pendingArticles: number;
  totalUsers: number;
  totalCategories: number;
  totalViews: number;
}

export interface CategoryStat {
  categoryName: string;
  count: number;
  /** @nullable */
  color?: string | null;
}

export interface ActivityItem {
  id: number;
  action: string;
  articleTitle: string;
  userName: string;
  createdAt: string;
}

export type ListArticlesParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
status?: ListArticlesStatus;
};

export type ListArticlesStatus = typeof ListArticlesStatus[keyof typeof ListArticlesStatus] | null;


export const ListArticlesStatus = {
  DRAFT: 'DRAFT',
  PENDING_REVIEW: 'PENDING_REVIEW',
  PUBLISHED: 'PUBLISHED',
} as const;