# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BALE BELEQ NEWS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/setup:
    post:
      operationId: createSuperAdmin
      tags: [auth]
      summary: Create first super admin (setup wizard)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupInput"
      responses:
        "201":
          description: Super admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
        "409":
          description: Setup already done
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Public articles
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List published articles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [DRAFT, PENDING_REVIEW, PUBLISHED, null]
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleList"

    post:
      operationId: createArticle
      tags: [articles]
      summary: Create article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
...
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found

    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /articles/{id}/publish:
    post:
      operationId: publishArticle
      tags: [articles]
      summary: Publish an article (Editor/Super Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}/submit:
    post:
      operationId: submitArticle
      tags: [articles]
      summary: Submit article for review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}/related:
    get:
      operationId: getRelatedArticles
      tags: [articles]
      summary: Get related articles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"

  # Categories
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Users
...
components:
  schemas:
...
    SetupInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, EDITOR, WARTAWAN]
        createdAt:
          type: string
...
    Article:
      type: object
      required: [id, title, slug, content, status, authorId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImage:
          type: ["string", "null"]
        status:
          type: string
          enum: [DRAFT, PENDING_REVIEW, PUBLISHED]
        isBreaking:
          type: boolean
        isFeatured:
          type: boolean
        authorId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        category:
          $ref: "#/components/schemas/Category"
        author:
          $ref: "#/components/schemas/User"
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        viewCount:
          type: integer
    ArticleList:
      type: object
      required: [articles, total, page, limit]
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ArticleInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImage:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        isBreaking:
          type: boolean
        isFeatured:
          type: boolean
        status:
          type: string
          enum: [DRAFT, PENDING_REVIEW, PUBLISHED]
    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImage:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        isBreaking:
          type: boolean
        isFeatured:
          type: boolean
        status:
          type: string
          enum: [DRAFT, PENDING_REVIEW, PUBLISHED]
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        articleCount:
          type: integer
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
...