# lib/api-zod/src/generated/types/articleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BALE BELEQ NEWS API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleUpdateStatus } from './articleUpdateStatus';

export interface ArticleUpdate {
  title?: string;
  /** @nullable */
  excerpt?: string | null;
  content?: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  categoryId?: number | null;
  isBreaking?: boolean;
  isFeatured?: boolean;
  status?: ArticleUpdateStatus;
}