# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BALE BELEQ NEWS API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}